refactor: 대표 장비를 등록하는 기능을 InventoryProduct에서 수행하도록 변경
