+++ backend/src/docs/asciidoc/inventoryProducts.adoc
@@ -1,6 +1,10 @@
 [[InventoryProduct]]
 == InventoryProduct API
 
+=== 나의 대표 장비 업데이트
+
+operation::inventoryProducts-update[snippets='http-request,http-response']
+
 === 나의 로그인 정보로 인벤토리 상품을 조회
 
 operation::inventoryProducts-get-my[snippets='http-request,http-response']
